threads

+++ src/gpu/tensor_kernels.cu
@@ -31,30 +31,30 @@
 #include <cuda.h>
 
 ///////////////////////////////////////////
-__global__ void sum_mat_row(float* a, float* b, float* c, int rows, int cols)
+__global__ void sum_mat_row(float* a, float* b, float* c, long int rows, long int cols)
 {
- int ops=rows*cols;
- int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
+ long int ops=rows*cols;
+ long int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
 
  if (thread_id_x < ops)
    c[thread_id_x]=a[thread_id_x]+b[thread_id_x%cols];
 
 }
 ///////////////////////////////////////////
-__global__ void sum_mat_col(float* a, float* b, float* c, int rows, int cols)
+__global__ void sum_mat_col(float* a, float* b, float* c, long int rows, long int cols)
 {
- int ops=rows*cols;
- int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
+ long int ops=rows*cols;
+ long int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
 
  if (thread_id_x < ops)
    c[thread_id_x]=a[thread_id_x]+b[thread_id_x/cols];
 
 }
 ///////////////////////////////////////////
-__global__ void set(float* a, float v, int rows, int cols)
+__global__ void set(float* a, float v, long int rows, long int cols)
 {
- int ops=rows*cols;
- int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
+ long int ops=rows*cols;
+ long int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
 
  if (thread_id_x < ops)
    a[thread_id_x]=v;
@@ -62,20 +62,20 @@ __global__ void set(float* a, float v, int rows, int cols)
 }
 
 ///////////////////////////////////////////
-__global__ void mult(float* a, float v, int rows, int cols)
+__global__ void mult(float* a, float v, long int rows, long int cols)
 {
- int ops=rows*cols;
- int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
+ long int ops=rows*cols;
+ long int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
 
  if (thread_id_x < ops)
    a[thread_id_x]*=v;
 
 }
 ///////////////////////////////////////////
-__global__ void el_mult(float* a, float *b, float *c, int incC, int rows, int cols)
+__global__ void el_mult(float* a, float *b, float *c, long int incC, long int rows, long int cols)
 {
-  int ops=rows*cols;
-  int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
+  long int ops=rows*cols;
+  long int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
 
   if (thread_id_x < ops)
     if (incC) c[thread_id_x]+=a[thread_id_x]*b[thread_id_x];
@@ -83,10 +83,10 @@ __global__ void el_mult(float* a, float *b, float *c, int incC, int rows, int co
 }
 
 ///////////////////////////////////////////
-__global__ void el_div(float* a, float *b, float *c, int incC, int rows, int cols)
+__global__ void el_div(float* a, float *b, float *c, long int incC, long int rows, long int cols)
 {
-  int ops=rows*cols;
-  int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
+  long int ops=rows*cols;
+  long int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
 
   if (thread_id_x < ops)
     if (incC) c[thread_id_x]+=a[thread_id_x]/b[thread_id_x];
@@ -94,19 +94,19 @@ __global__ void el_div(float* a, float *b, float *c, int incC, int rows, int col
 }
 
 ///////////////////////////////////////////
-__global__ void sum(float* a, float v, int rows, int cols)
+__global__ void sum(float* a, float v, long int rows, long int cols)
 {
- int ops=rows*cols;
- int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
+ long int ops=rows*cols;
+ long int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
 
  if (thread_id_x < ops)
    a[thread_id_x]+=v;
 
 }
 ///////////////////////////////////////////
-__global__ void sum(float scA,float* a,float scB,float *b, float *c,int incC, int tam)
+__global__ void sum(float scA,float* a,float scB,float *b, float *c,long int incC, long int tam)
 {
-  int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
+  long int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
 
   if (thread_id_x < tam) {
     if (incC) c[thread_id_x]+=scA*a[thread_id_x]+scB*b[thread_id_x];
@@ -115,12 +115,12 @@ __global__ void sum(float scA,float* a,float scB,float *b, float *c,int incC, in
 }
 
 ///////////////////////////////////////////
-__global__ void reduce_array_sum(float* array, long int ops, int cols,float* result)
+__global__ void reduce_array_sum(float* array, long int ops, long int cols,float* result)
 {
   extern __shared__ float arr_acc[];
   __shared__ float accumulate_result[1];
 
-  int thread_id_x = threadIdx.x +blockIdx.x*blockDim.x;
+  long int thread_id_x = threadIdx.x +blockIdx.x*blockDim.x;
   float sum=0;
   arr_acc[thread_id_x]=0.0;
 
@@ -130,7 +130,7 @@ __global__ void reduce_array_sum(float* array, long int ops, int cols,float* res
   __syncthreads();
   if (thread_id_x<ops)
   {
-  	for (int i=0; i<cols;i++)
+  	for (long int i=0; i<cols;i++)
     		sum+=array[thread_id_x*cols+i];
 
   __syncthreads();
@@ -141,7 +141,7 @@ __global__ void reduce_array_sum(float* array, long int ops, int cols,float* res
 
   if (thread_id_x==0)
   {
-  	for (int i=0; i<ops;i++)
+  	for (long int i=0; i<ops;i++)
       accumulate_result[thread_id_x]+=arr_acc[thread_id_x+i];
 
     result[thread_id_x]=accumulate_result[thread_id_x];//copy back to global memory from shared
@@ -150,10 +150,10 @@ __global__ void reduce_array_sum(float* array, long int ops, int cols,float* res
 }
 
 ///////////////////////////////////////////
-__global__ void log(float* a, int rows, int cols)
+__global__ void log(float* a, long int rows, long int cols)
 {
- int ops=rows*cols;
- int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
+ long int ops=rows*cols;
+ long int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
 
  if (thread_id_x < ops)
    a[thread_id_x]=log(a[thread_id_x]);
@@ -161,10 +161,10 @@ __global__ void log(float* a, int rows, int cols)
 }
 
 ///////////////////////////////////////////
-__global__ void exp(float* a, int rows, int cols)
+__global__ void exp(float* a, long int rows, long int cols)
 {
- int ops=rows*cols;
- int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
+ long int ops=rows*cols;
+ long int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
 
  if (thread_id_x < ops)
    a[thread_id_x]=exp(a[thread_id_x]);
@@ -172,10 +172,10 @@ __global__ void exp(float* a, int rows, int cols)
 }
 
 ///////////////////////////////////////////
-__global__ void sqrt(float* a, int rows, int cols)
+__global__ void sqrt(float* a, long int rows, long int cols)
 {
- int ops=rows*cols;
- int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
+ long int ops=rows*cols;
+ long int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
 
  if (thread_id_x < ops)
    a[thread_id_x]=sqrt(a[thread_id_x]);
@@ -183,10 +183,10 @@ __global__ void sqrt(float* a, int rows, int cols)
 }
 
 ///////////////////////////////////////////
-__global__ void sqr(float* a, int rows, int cols)
+__global__ void sqr(float* a, long int rows, long int cols)
 {
- int ops=rows*cols;
- int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
+ long int ops=rows*cols;
+ long int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
 
  if (thread_id_x < ops)
    a[thread_id_x]*=a[thread_id_x];
@@ -194,10 +194,10 @@ __global__ void sqr(float* a, int rows, int cols)
 }
 
 ///////////////////////////////////////////
-__global__ void mask(float* a, float v, int rows, int cols)
+__global__ void mask(float* a, float v, long int rows, long int cols)
 {
- int ops=rows*cols;
- int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
+ long int ops=rows*cols;
+ long int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
 
  if (thread_id_x < ops)
    a[thread_id_x]=a[thread_id_x]<v;
@@ -206,10 +206,10 @@ __global__ void mask(float* a, float v, int rows, int cols)
 
 ///////////////////////////////////////////
 
-__global__ void reduce_sum2D(float *a,float *b,int rows,int cols,int axis)
+__global__ void reduce_sum2D(float *a,float *b,long int rows,long int cols,long int axis)
 {
-  int ops=rows*cols;
-  int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
+  long int ops=rows*cols;
+  long int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
 
   if (thread_id_x < ops)
     if (axis==0)
@@ -218,10 +218,10 @@ __global__ void reduce_sum2D(float *a,float *b,int rows,int cols,int axis)
         b[thread_id_x/cols]+=a[thread_id_x];
 }
 ///////////////////////////////////////////
-__global__ void cent(float* a, float* b, float* c, int tam)
+__global__ void cent(float* a, float* b, float* c, long int tam)
 {
 
- int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
+ long int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
 
  if (thread_id_x < tam){
    c[thread_id_x]=0;
@@ -231,21 +231,21 @@ __global__ void cent(float* a, float* b, float* c, int tam)
 }
 
 
-__global__ void accuracy(float* T, float* N,float* acc,int cols, long int total_ops, int* MC_err)
+__global__ void accuracy(float* T, float* N,float* acc,long int cols, long int total_ops, int* MC_err)
 {
 
-int thread_id_x = threadIdx.x + blockIdx.x*blockDim.x;
-int result_t=T[thread_id_x*cols];
+long int thread_id_x = threadIdx.x + blockIdx.x*blockDim.x;
+long int result_t=T[thread_id_x*cols];
 float result_n=N[thread_id_x*cols];
 
-int row_max_t=0;
-int row_max_n=0;
+long int row_max_t=0;
+long int row_max_n=0;
 
-int aux_t;
+long int aux_t;
 float aux_n;
 if (thread_id_x < total_ops)
 {
-  for(int i=1;i<cols;i++)
+  for(long int i=1;i<cols;i++)
   {
    aux_t=T[thread_id_x*cols+i];
    aux_n=N[thread_id_x*cols+i];
@@ -263,15 +263,15 @@ if (thread_id_x < total_ops)
   }
 
   acc[thread_id_x]=row_max_t;
-  atomicAdd(MC_err,(int)(row_max_t==row_max_n));
+  atomicAdd(MC_err,(long int)(row_max_t==row_max_n));
 }
 
 }
 
 ///////////////////////////////////////////
-__global__ void relu(float *a,float *b,int tam)
+__global__ void relu(float *a,float *b,long int tam)
 {
-  int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
+  long int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
 
   if (thread_id_x < tam){
     if (a[thread_id_x]>0.0) b[thread_id_x]=a[thread_id_x];
@@ -280,9 +280,9 @@ __global__ void relu(float *a,float *b,int tam)
 }
 
 
-__global__ void d_relu(float *d,float *i,float *pd,int tam)
+__global__ void d_relu(float *d,float *i,float *pd,long int tam)
 {
-  int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
+  long int thread_id_x = threadIdx.x+blockIdx.x*blockDim.x;
 
   if (thread_id_x < tam){
     if (i[thread_id_x]>0.0) pd[thread_id_x]=d[thread_id_x];
@@ -295,24 +295,24 @@ __global__ void d_relu(float *d,float *i,float *pd,int tam)
 __global__ void softmax(float* E,float* N,float* auxE ,long int sample_dim, long int n_vals)
 {
     float C_value=0;
-    int thread_id_x = threadIdx.x + blockIdx.x*blockDim.x;
+    long int thread_id_x = threadIdx.x + blockIdx.x*blockDim.x;
     float maxCoef = E[thread_id_x*sample_dim];
     float actualCoef = 0;
     if (thread_id_x<n_vals)
     {
 
-	    for (int cA = 1; cA < sample_dim; cA++)
+	    for (long int cA = 1; cA < sample_dim; cA++)
     		if (E[thread_id_x*sample_dim+cA] > maxCoef)
     			 maxCoef=E[thread_id_x*sample_dim+cA];
 
-	    for (int cA = 0; cA < sample_dim; cA++)
+	    for (long int cA = 0; cA < sample_dim; cA++)
   		{
   			actualCoef=expf(E[thread_id_x*sample_dim+cA]-maxCoef);
   			auxE[thread_id_x*sample_dim+cA]=actualCoef;
         C_value+=actualCoef;
   		}
 
-      for (int cA=0; cA < sample_dim; cA++)
+      for (long int cA=0; cA < sample_dim; cA++)
 	       N[thread_id_x*sample_dim+cA]=auxE[thread_id_x*sample_dim+cA]/C_value;
     }
