+++ src/gpu/tensor_kernels.h
@@ -29,33 +29,33 @@
 
 #include <cuda.h>
 
-__global__ void set(float* a, float v, int rows, int cols);
-__global__ void mult(float* a, float v, int rows, int cols);
-__global__ void sum(float* a, float v, int rows, int cols);
+__global__ void set(float* a, float v, long int rows, long int cols);
+__global__ void mult(float* a, float v, long int rows, long int cols);
+__global__ void sum(float* a, float v, long int rows, long int cols);
 
 
-__global__ void log(float* a, int rows, int cols);
-__global__ void exp(float* a, int rows, int cols);
-__global__ void sqrt(float* a,int rows, int cols);
-__global__ void sqr(float* a, int rows, int cols);
-__global__ void mask(float* a, float v, int rows, int cols);
+__global__ void log(float* a, long int rows, long int cols);
+__global__ void exp(float* a, long int rows, long int cols);
+__global__ void sqrt(float* a,long int rows, long int cols);
+__global__ void sqr(float* a, long int rows, long int cols);
+__global__ void mask(float* a, float v, long int rows, long int cols);
 
-__global__ void reduce_array_sum(float* array, long int ops, int cols,float* result);
+__global__ void reduce_array_sum(float* array, long int ops, long int cols,float* result);
 
-__global__ void sum(float scA,float* a,float scB,float *b, float *c,int incC, int tam);
-__global__ void el_mult(float* a,float *b, float *c, int incC, int rows, int cols);
-__global__ void el_div(float* a, float *b, float *c, int incC, int rows, int cols);
+__global__ void sum(float scA,float* a,float scB,float *b, float *c,long int incC, long int tam);
+__global__ void el_mult(float* a,float *b, float *c, long int incC, long int rows, long int cols);
+__global__ void el_div(float* a, float *b, float *c, long int incC, long int rows, long int cols);
 
-__global__ void sum_mat_row(float* a, float* b, float* c, int cols, int rows);
-__global__ void sum_mat_col(float* a, float* b, float* c, int cols, int rows);
+__global__ void sum_mat_row(float* a, float* b, float* c, long int cols, long int rows);
+__global__ void sum_mat_col(float* a, float* b, float* c, long int cols, long int rows);
 
-__global__ void reduce_sum2D(float *a,float *b,int r,int c,int axis);
+__global__ void reduce_sum2D(float *a,float *b,long int r,long int c,long int axis);
 
-__global__ void cent(float* a, float* b, float* c, int tam);
-__global__ void accuracy(float* T, float* N,float* acc,int cols, long int total_ops, int* MC_err);
+__global__ void cent(float* a, float* b, float* c, long int tam);
+__global__ void accuracy(float* T, float* N,float* acc,long int cols, long int total_ops, int* MC_err);
 
-__global__ void relu(float *a,float *b,int tam);
-__global__ void d_relu(float *d,float *i,float *pd,int tam);
+__global__ void relu(float *a,float *b,long int tam);
+__global__ void d_relu(float *d,float *i,float *pd,long int tam);
 __global__ void softmax(float* E,float* N,float* auxE ,long int sample_dim, long int n_vals);
 
 #endif
